backend/ai/requirements.txt
torch
numpy
neo4j
fasttext
nltk
python-docx
pdfplumber